# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget API",
        "description": "Documentación para Budget API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "admin@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/egresses": {
            "get": {
                "summary": "List all egresses",
                "operationId": "89417902769bed0e4c328319d689467d",
                "responses": {
                    "200": {
                        "description": "A list of egresses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Store a new egress",
                "operationId": "8a1ec8a43b49f0e7eab79075152f128d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "amount",
                                    "is_fixed",
                                    "date_egreso",
                                    "status",
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "is_fixed": {
                                        "type": "boolean"
                                    },
                                    "date_egreso": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "payment_method_id": {
                                        "type": "integer"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Egress created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/egresses/{id}": {
            "get": {
                "summary": "Show a specific egress",
                "operationId": "61120d77b039be6215a71ba636bad851",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Egress details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a specific egress",
                "operationId": "35e30ef4889692bfc9c28f4dbf420629",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "amount",
                                    "is_fixed",
                                    "date_egreso",
                                    "status",
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "is_fixed": {
                                        "type": "boolean"
                                    },
                                    "date_egreso": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "payment_method_id": {
                                        "type": "integer"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Egress updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a specific egress",
                "operationId": "7e72abcfcd73bda7107eb87bf26a0343",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Egress deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/top-categories": {
            "get": {
                "summary": "Get top categories",
                "operationId": "0797a675c2b97a06d5e2365061574806",
                "responses": {
                    "200": {
                        "description": "Top categories with total spent and remaining budget"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories-with-spent": {
            "get": {
                "summary": "Get all categories with spent amount",
                "operationId": "8e20fe430865ae46d77bfeb94b7b7482",
                "responses": {
                    "200": {
                        "description": "List of categories with spent amount"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cuentas-por-pagar": {
            "get": {
                "summary": "Get cuentas por pagar",
                "operationId": "ad8f0cd4cfb9d23b7c1c8a7dbfd2c24a",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cuentas por pagar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/export-cuentas-por-pagar": {
            "get": {
                "summary": "Export cuentas por pagar",
                "operationId": "d4d45eedb6ad36dcac5331a7de26fc3c",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported file"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/export-payment-method-view": {
            "get": {
                "summary": "Export payment method view",
                "operationId": "03b6d522901817aa51f8c9942ce7d6c5",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported file"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-egresses": {
            "get": {
                "summary": "Get list of egresses",
                "operationId": "0d3bcc106cd1c3915a3baed67282198b",
                "responses": {
                    "200": {
                        "description": "List of egresses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/spending-by-payment-method": {
            "get": {
                "summary": "Get spending by payment method",
                "operationId": "878923587b25275e236ac4a0c9b9bd21",
                "parameters": [
                    {
                        "name": "payment_method",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spending by payment method"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}